pyproject.toml
[project]
name = "quitanda-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.128.0",
    "pyjwt>=2.11.0",
    "python-slugify[unidecode]>=8.0.4",
    "sqlalchemy>=2.0.46",
]

